+++ e-commerce/src/components/AutoSlide/Arrows.jsx
