+++ e-commerce/src/components/AutoSlide/AutoSlide.jsx
@@ -1,0 +1,43 @@
+import React, { useEffect, useState } from "react";
+import SliderContent from "./SliderContent";
+import Dots from "./Dots";
+// import Arrows from "./Arrows";
+import sliderImage from "./sliderImage";
+import "./slider.css";
+
+const len = sliderImage.length - 1;
+
+function AutoSlide(props) {
+  const [activeIndex, setActiveIndex] = useState(0);
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setActiveIndex(activeIndex === len ? 0 : activeIndex + 1);
+    }, 3000);
+    return () => clearInterval(interval);
+  }, [activeIndex]);
+
+  return (
+    <div className="d-flex bg_sky_blue">
+      <div className="slider-container my-4">
+        <SliderContent activeIndex={activeIndex} sliderImage={sliderImage} />
+        <Dots
+          activeIndex={activeIndex}
+          sliderImage={sliderImage}
+          onclick={(activeIndex) => setActiveIndex(activeIndex)}
+        />
+      </div>
+
+      <div className="slider-container my-4">
+        <SliderContent activeIndex={activeIndex} sliderImage={sliderImage} />
+        <Dots
+          activeIndex={activeIndex}
+          sliderImage={sliderImage}
+          onclick={(activeIndex) => setActiveIndex(activeIndex)}
+        />
+      </div>
+    </div>
+  );
+}
+
+export default AutoSlide;
