+++ e-commerce/src/components/AutoSlide/SliderContent.jsx
@@ -6,7 +6,7 @@ function SliderContent({ activeIndex, sliderImage }) {
       {sliderImage.map((slide, index) => (
         <div
           key={index}
-          className={`${index === activeIndex ? "slides active" : "inactive"} bg-light`}
+          className={`${index === activeIndex ? "slides active" : "inactive"} bg-white`}
           style={{background: "black", color: "white"}}
         >
           {/* <img className="slide-image" src={slide.urls} alt="" /> */}
