+++ e-commerce/src/components/AutoSlide/slider.css
@@ -1,3 +1,6 @@
+.bg_sky_blue {
+  background-color: var(--sky_blue);
+}
 
 :root {
   --heights: 30vh;
@@ -25,12 +28,13 @@
   height: var(--heights);
   width: var(--widths);
   position: relative;
-  border-radius: 15px;
+  border-radius: 10px;
 }
 
 .slide-image {
   width: 100%;
   height: 100%;
+  border-radius: 10px;
   /* position: absolute; */
   object-fit: cover;
 }
