+++ e-commerce/src/components/AutoSlide/sliderImage.jsx
