+++ e-commerce/src/index.css
@@ -3,6 +3,7 @@
 
 :root {
   --primary_color: #FF7E07;
+  --sky_blue: #D1EFCD
 }
 
 body {
